Added Art, Fixed Sync, Created Dev Server
Added AI-Gen art for the cards and modified the cards to use these images.

Fixed Issues with card selection syncing and with in game selections syncing.

Created Dev Server Build along with pool, and que for matchmaking so that anyone can play the 'release' build while I am working on the dev build